refactor
#52
