+++ src/app/containers/config-editor/config-editor.component.html
@@ -7,24 +7,7 @@
     </button>
   </mat-toolbar-row>
   <mat-toolbar-row>
-    <button
-      (click)="fileInput.click()"
-      mat-flat-button
-      color="primary"
-      class="m-r1"
-    >
-      Load Styles
-    </button>
-    <input
-      hidden
-      (change)="onFileSelected(fileInput)"
-      #fileInput
-      type="file"
-      id="file"
-    />
-    <button (click)="saveStyles()" mat-flat-button color="primary">
-      Save Styles
-    </button>
+    <app-style-actions [layers]="layers$ | async"></app-style-actions>
   </mat-toolbar-row>
 </mat-toolbar>
