+++ src/app/containers/config-editor/config-editor.component.ts
@@ -13,32 +13,7 @@ import { HereTileLayer } from 'src/app/models/here-tile-layer.model';
 export class ConfigEditorComponent {
   @Output() toggleNav = new EventEmitter();
 
-  private layers: HereTileLayer[];
+  layers$ = this.store.pipe(select(fromStore.getLayers));
 
-  constructor(
-    private store: Store<fromStore.State>,
-    private _FileSaverService: FileSaverService
-  ) {
-    this.store
-      .pipe(select(fromStore.getLayers))
-      .subscribe(layers => (this.layers = layers));
-  }
-
-  loadMapStyle(value: string) {
-    const layers: HereTileLayer[] = JSON.parse(value);
-    this.store.dispatch(fromStore.loadLayers({ layers }));
-  }
-
-  onFileSelected(fileInput: HTMLInputElement) {
-    if (fileInput.files && fileInput.files.length) {
-      const fileReader = new FileReader();
-      fileReader.readAsText(fileInput.files.item(0));
-      fileReader.onload = () => this.loadMapStyle(fileReader.result as string);
-      fileInput.value = '';
-    }
-  }
-
-  saveStyles() {
-    this._FileSaverService.saveText(JSON.stringify(this.layers), 'styles.json');
-  }
+  constructor(private store: Store<fromStore.State>) {}
 }
