+++ src/app/containers/style-actions/style-actions.component.html
@@ -1,0 +1,18 @@
+<button
+  (click)="fileInput.click()"
+  mat-flat-button
+  color="primary"
+  class="m-r1"
+>
+  Load Styles
+</button>
+<input
+  hidden
+  (change)="onFileSelected(fileInput)"
+  #fileInput
+  type="file"
+  id="file"
+/>
+<button (click)="saveStyles()" mat-flat-button color="primary">
+  Save Styles
+</button>
